--- backend/circuit_generator.py
@@ -1,883 +1,0 @@
-#!/usr/bin/env python3
-"""
-Fixed Circuit Generator with External Expected Result
-
-This script generates quantum circuits from optimized MLIR and reads the
-expected result from an external file (expected_res.txt) instead of calculating it.
-
-Usage: python circuit_generator_fixed.py <optimized.mlir> <output.py> [expected_res.txt]
-"""
-
-import re
-import sys
-import os
-from typing import Dict, List, Tuple, Optional, Set, Any
-from dataclasses import dataclass
-import json
-
-
-def convert_4bit_to_signed(unsigned_value):
-    """
-    Convert 4-bit unsigned quantum result to signed interpretation
-    
-    4-bit signed range: -8 to +7
-    - Values 0-7: remain positive (0 to +7)
-    - Values 8-15: convert to negative (-8 to -1)
-    """
-    if unsigned_value >= 8:
-        return unsigned_value - 16  # Convert to negative
-    else:
-        return unsigned_value       # Already positive
-
-@dataclass
-class QubitRegister:
-    name: str
-    size: int
-    qiskit_name: str
-    initial_value: Optional[int] = None
-
-@dataclass
-class QuantumOperation:
-    op_type: str
-    operands: List[str]
-    result: Optional[str] = None
-    optimization: str = ""
-    description: str = ""
-    attributes: Dict[str, Any] = None
-
-    def __post_init__(self):
-        if self.attributes is None:
-            self.attributes = {}
-
-class FixedQiskitGenerator:
-    def __init__(self):
-        self.registers: Dict[str, QubitRegister] = {}
-        self.operations: List[QuantumOperation] = []
-        self.classical_registers: Dict[str, str] = {}
-        self.optimizations_applied: List[str] = []
-        self.initial_values: Dict[str, int] = {}
-        self.expected_result: Optional[int] = None
-        
-    def load_expected_result(self, expected_file: str = "expected_res.txt") -> int:
-        """Load expected result from external file"""
-        try:
-            if os.path.exists(expected_file):
-                with open(expected_file, 'r') as f:
-                    expected = int(f.read().strip())
-                print(f"📊 Loaded expected result from {expected_file}: {expected}")
-                return expected
-            else:
-                print(f"⚠️  Expected result file {expected_file} not found, using default: 0")
-                return 0
-        except Exception as e:
-            print(f"❌ Error loading expected result: {e}")
-            return 0
-        
-    def convert_4bit_to_signed(unsigned_value):
-        """Convert 4-bit unsigned to signed interpretation"""
-        if unsigned_value >= 8:
-            return unsigned_value - 16
-        else:
-            return unsigned_value
-    
-    def parse_optimized_mlir(self, content: str) -> None:
-        """Parse any optimized MLIR content dynamically"""
-        lines = content.split('\n')
-        
-        # Try to extract expected result from comment first
-        expected_from_comment = None
-        for line in lines:
-            if line.startswith("// Expected classical result:"):
-                try:
-                    expected_from_comment = int(line.split(":")[-1].strip())
-                    print(f"📊 Found expected result from MLIR comment: {expected_from_comment}")
-                    self.expected_result = expected_from_comment
-                except:
-                    pass
-                break
-        
-        # Extract applied optimizations from header
-        for line in lines:
-            if line.startswith("// Applied optimizations:"):
-                opt_line = line.replace("// Applied optimizations:", "").strip()
-                self.optimizations_applied = [opt.strip() for opt in opt_line.split(",")]
-                break
-        
-        parsed_lines = 0
-        for i, line in enumerate(lines):
-            line = line.strip()
-            
-            # Skip empty and comment lines
-            if not line or line.startswith('//'):
-                continue
-                
-            # Skip structural MLIR elements
-            if any(skip in line for skip in ['builtin.module', 'quantum.func', 'func.return', 'func_name']):
-                continue
-                
-            if line == '}' or line == '{':
-                continue
-                
-            print(f"🔍 Processing line {i+1}: {line}")
-                
-            try:
-                if self._parse_line(line):
-                    parsed_lines += 1
-                    print(f"   ✅ Successfully parsed")
-                else:
-                    print(f"   ⚠️  No match found")
-            except Exception as e:
-                print(f"   ❌ Error: {e}")
-        
-        print(f"\n📊 Parsing Summary:")
-        print(f"   Total lines processed: {len(lines)}")
-        print(f"   Lines successfully parsed: {parsed_lines}")
-        print(f"   Registers found: {len(self.registers)}")
-        print(f"   Operations found: {len(self.operations)}")
-        
-        if self.registers:
-            print(f"📋 Registers: {list(self.registers.keys())}")
-        if self.optimizations_applied:
-            print(f"🔧 Optimizations: {self.optimizations_applied}")
-        if self.operations:
-            print(f"🔬 Operations found:")
-            for i, op in enumerate(self.operations):
-                print(f"   {i+1}. {op.op_type}: {op.operands}")
-        print()
-    
-    def _parse_line(self, line: str) -> bool:
-        """Parse a single MLIR line with all operation types - returns True if parsed successfully"""
-        
-        # Parse allocation: %q0 = q.alloc : !qreg<4>
-        alloc_match = re.search(r'%(\w+)\s*=\s*q\.alloc\s*:\s*!qreg<(\d+)>', line)
-        if alloc_match:
-            reg_name, size = alloc_match.groups()
-            full_reg_name = f"%{reg_name}"
-            
-            if full_reg_name in self.registers:
-                return True
-            
-            qiskit_name = f"q{len(self.registers)}"
-            self.registers[full_reg_name] = QubitRegister(
-                name=full_reg_name,
-                size=int(size),
-                qiskit_name=qiskit_name
-            )
-            print(f"   📦 Allocated register: {full_reg_name} -> {qiskit_name} ({size} qubits)")
-            return True
-        
-        # Parse initialization: q.init %q0, 1 : i32
-        init_match = re.search(r'q\.init\s+%(\w+),\s*(\d+)', line)
-        if init_match:
-            reg_name, value = init_match.groups()
-            full_reg_name = f"%{reg_name}"
-            int_value = int(value)
-            
-            self.initial_values[full_reg_name] = int_value
-            if full_reg_name in self.registers:
-                self.registers[full_reg_name].initial_value = int_value
-            
-            self.operations.append(QuantumOperation(
-                op_type="init",
-                operands=[full_reg_name],
-                attributes={"value": int_value},
-                description=f"Initialize {full_reg_name} = {int_value}"
-            ))
-            print(f"   🔧 Initialize: {full_reg_name} = {int_value}")
-            return True
-        
-        # Parse circuit operations: q.add_circuit %q0, %q1, %q2
-        circuit_match = re.search(r'q\.(\w+_circuit)\s+((?:%\w+(?:,\s*)?)+)(?:\s*//\s*(.+))?', line)
-        if circuit_match:
-            circuit_type, operands_str, annotation = circuit_match.groups()
-            operands = [op.strip() for op in operands_str.split(',')]
-            
-            self.operations.append(QuantumOperation(
-                op_type=circuit_type,
-                operands=operands,
-                result=operands[-1] if len(operands) > 2 else None,
-                optimization=annotation or "",
-                description=f"{circuit_type} operation"
-            ))
-            print(f"   ⚡ Circuit: {circuit_type} with {operands}")
-            return True
-        
-        # Parse measurement: %q3 = q.measure %q2 : !qreg -> i32
-        measure_match = re.search(r'%(\w+)\s*=\s*q\.measure\s+%(\w+)', line)
-        if measure_match:
-            result, operand = measure_match.groups()
-            result_name = f"%{result}"
-            operand_name = f"%{operand}"
-            
-            classical_reg_name = f"c{len(self.classical_registers)}"
-            self.classical_registers[result_name] = classical_reg_name
-            
-            self.operations.append(QuantumOperation(
-                op_type="measure",
-                operands=[operand_name],
-                result=result_name,
-                description=f"Measure {operand_name}"
-            ))
-            print(f"   📏 Measure: {operand_name} -> {result_name}")
-            return True
-        
-        # Parse basic gates: q.cx %q0[0], %q1[0]
-        gate_match = re.search(r'q\.(\w+)\s+((?:%\w+(?:\[\d+\])?(?:,\s*)?)+)(?:\s*//\s*(.+))?', line)
-        if gate_match and not gate_match.group(1).endswith('_circuit') and gate_match.group(1) != 'measure':
-            gate_type, operands_str, annotation = gate_match.groups()
-            
-            # Skip comment pseudo-operations
-            if gate_type == 'comment':
-                return True
-                
-            operands = [op.strip() for op in operands_str.split(',')]
-            
-            # Extract all register names from operands to ensure they're tracked
-            for operand in operands:
-                if '[' in operand:
-                    reg_name = operand.split('[')[0]
-                else:
-                    reg_name = operand
-                
-                # Ensure this register exists in our tracking
-                if reg_name.startswith('%') and reg_name not in self.registers:
-                    # Auto-create register if it doesn't exist (from decomposed operations)
-                    print(f"   🔍 Auto-creating register {reg_name} found in gate operation")
-                    qiskit_name = f"q{len(self.registers)}"
-                    self.registers[reg_name] = QubitRegister(
-                        name=reg_name,
-                        size=4,  # Default to 4-bit registers
-                        qiskit_name=qiskit_name
-                    )
-            
-            self.operations.append(QuantumOperation(
-                op_type=gate_type,
-                operands=operands,
-                optimization=annotation or "",
-                description=f"{gate_type.upper()} gate"
-            ))
-            print(f"   🚪 Gate: {gate_type} on {operands}")
-            return True
-            
-        # If we get here, no pattern matched
-        return False
-    
-    def _detect_operation_from_gates(self):
-        """
-        Detect the original C operation by analyzing the quantum gate pattern
-        """
-        # Count different gate types
-        gate_counts = {}
-        total_gates = 0
-        
-        for op in self.operations:
-            if op.op_type in ['x', 'cx', 'ccx', 'measure']:
-                gate_counts[op.op_type] = gate_counts.get(op.op_type, 0) + 1
-                if op.op_type != 'measure':
-                    total_gates += 1
-        
-        print(f"🔧 Gate analysis: {gate_counts}, total: {total_gates}")
-        
-        # Heuristics based on quantum gate patterns:
-        ccx_count = gate_counts.get('ccx', 0)
-        cx_count = gate_counts.get('cx', 0)
-        x_count = gate_counts.get('x', 0)
-        
-        # Division: Complex pattern with many CCX gates (20+ gates)
-        if ccx_count >= 15 and total_gates >= 70:
-            return "div"
-        
-        # Modulo: Medium complexity (10-20 CCX gates)
-        elif ccx_count >= 8 and ccx_count < 15:
-            return "mod"
-        
-        # Multiplication: Medium complexity with specific pattern
-        elif ccx_count >= 4 and ccx_count < 12 and total_gates >= 20:
-            return "mul"
-        
-        # Bitwise AND: Simple CCX pattern (2-4 CCX gates)
-        elif ccx_count >= 2 and ccx_count <= 4 and cx_count <= 2:
-            return "and"
-        
-        # Bitwise OR: More CX than CCX gates
-        elif cx_count > ccx_count and ccx_count >= 1:
-            return "or"
-        
-        # XOR: Mainly CX gates, few CCX
-        elif cx_count >= 4 and ccx_count <= 2:
-            return "xor"
-        
-        # NOT: Many X gates with CX gates
-        elif x_count >= 4 and cx_count >= 4:
-            return "not"
-        
-        # Addition: Balanced CX and CCX
-        elif cx_count >= 4 and ccx_count >= 2 and ccx_count <= 6:
-            return "add"
-        
-        # Subtraction: Similar to addition but different pattern
-        elif cx_count >= 2 and ccx_count >= 1 and total_gates <= 15:
-            return "sub"
-        
-        # Negation: Few gates, mainly X
-        elif x_count >= 2 and total_gates <= 10:
-            return "neg"
-        
-        # Logical operations (&&, ||): Usually result in 0 or 1
-        elif total_gates <= 20 and ccx_count <= 6:
-            # Check if this might be logical vs bitwise
-            return "logical_and"  # Will be handled specially
-        
-        # Default fallback
-        else:
-            print(f"⚠️ Unknown pattern: CCX={ccx_count}, CX={cx_count}, X={x_count}")
-            return "unknown"
-
-    def _calculate_binary_operation(self, operation, a, b):
-        """Calculate result for binary operations"""
-        print(f"📊 Binary operation: {operation}({a}, {b})")
-        
-        if operation == "div":
-            result = a // b if b != 0 else 0
-            print(f"   Division: {a} ÷ {b} = {result}")
-            
-        elif operation == "mod":
-            result = a % b if b != 0 else 0
-            print(f"   Modulo: {a} % {b} = {result}")
-            
-        elif operation == "mul":
-            result = (a * b) & 0xF  # 4-bit mask
-            print(f"   Multiplication: {a} × {b} = {result}")
-            
-        elif operation == "add":
-            result = (a + b) & 0xF
-            print(f"   Addition: {a} + {b} = {result}")
-            
-        elif operation == "sub":
-            result = (a - b) & 0xF
-            print(f"   Subtraction: {a} - {b} = {result}")
-            
-        elif operation == "and":
-            result = (a & b) & 0xF  # Bitwise AND
-            print(f"   Bitwise AND: {a} & {b} = {result}")
-            
-        elif operation == "or":
-            result = (a | b) & 0xF  # Bitwise OR
-            print(f"   Bitwise OR: {a} | {b} = {result}")
-            
-        elif operation == "xor":
-            result = (a ^ b) & 0xF
-            print(f"   XOR: {a} ^ {b} = {result}")
-            
-        elif operation == "logical_and":
-            result = 1 if (a != 0 and b != 0) else 0  # Logical AND (&&)
-            print(f"   Logical AND: {a} && {b} = {result}")
-            
-        elif operation == "logical_or":
-            result = 1 if (a != 0 or b != 0) else 0  # Logical OR (||)
-            print(f"   Logical OR: {a} || {b} = {result}")
-            
-        else:
-            print(f"   ❓ Unknown binary operation: {operation}")
-            result = a  # Fallback to first operand
-        
-        return result
-
-    def _calculate_unary_operation(self, operation, a):
-        """Calculate result for unary operations"""
-        print(f"📊 Unary operation: {operation}({a})")
-        
-        if operation == "not":
-            result = (~a) & 0xF  # 4-bit bitwise NOT
-            print(f"   Bitwise NOT: ~{a} = {result}")
-            print(f"   Binary: ~{a:04b} = {result:04b}")
-            
-        elif operation == "neg":
-            result = (-a) & 0xF  # 4-bit negation
-            print(f"   Negation: -{a} = {result}")
-            
-        elif operation == "logical_not":
-            result = 1 if a == 0 else 0  # Logical NOT (!)
-            print(f"   Logical NOT: !{a} = {result}")
-            
-        else:
-            print(f"   ❓ Unknown unary operation: {operation}")
-            result = a  # Fallback to input
-        
-        return result
-    
-    def _safe_get_register(self, reg_name: str) -> Optional[QubitRegister]:
-        """Safely get a register"""
-        if reg_name in self.registers:
-            return self.registers[reg_name]
-        
-        # Create fallback if needed
-        qiskit_name = f"fallback_q{len(self.registers)}"
-        fallback_reg = QubitRegister(name=reg_name, size=4, qiskit_name=qiskit_name)
-        self.registers[reg_name] = fallback_reg
-        return fallback_reg
-        
-    def _extract_qubit_index(self, operand: str) -> Tuple[str, Optional[int]]:
-        """Extract register name and qubit index from operand"""
-        if '[' in operand:
-            reg_part, index_part = operand.split('[', 1)
-            index = int(index_part.rstrip(']'))
-            return reg_part, index
-        return operand, None
-    
-    def generate_qiskit_code(self, expected_result: int) -> List[str]:
-        """Generate complete Qiskit Python code with external expected result"""
-        
-        lines = [
-            "#!/usr/bin/env python3",
-            "'''",
-            "Generated Qiskit Circuit from Optimized MLIR",
-            "",
-            f"Expected classical result: {expected_result}",
-            "'''",
-            "",
-            "from qiskit import QuantumCircuit, QuantumRegister, ClassicalRegister",
-            "from qiskit_aer import AerSimulator",
-            "from qiskit.visualization import plot_histogram",
-            "import matplotlib.pyplot as plt",
-            "",
-            "def convert_4bit_to_signed(unsigned_value):",
-            "    '''Convert 4-bit unsigned to signed interpretation'''",
-            "    if unsigned_value >= 8:",
-            "        return unsigned_value - 16",
-            "    else:",
-            "        return unsigned_value",
-            "",
-            "def initialize_register(qc, qreg, value, num_bits):",
-            "    '''Initialize quantum register to classical value'''",
-            "    bin_val = format(value, f'0{num_bits}b')[::-1]  # LSB first",
-            "    for i, bit in enumerate(bin_val):",
-            "        if bit == '1':",
-            "            qc.x(qreg[i])",
-            "    print(f'   Initialized {qreg.name} to {value} (binary: {bin_val[::-1]})')",
-            "",
-            "def apply_and_circuit(qc, a_reg, b_reg, result_reg):",
-            "    '''CORRECTED: Pure bitwise AND circuit using only necessary Toffoli gates'''",
-            "    print(f'   Applying pure AND circuit: {a_reg.name} & {b_reg.name} -> {result_reg.name}')",
-            "    # For bitwise AND: result[i] = a[i] & b[i] using Toffoli gates",
-            "    # Only add Toffoli gate if both input bits could be 1",
-            "    for i in range(min(len(a_reg), len(b_reg), len(result_reg))):",
-            "        qc.ccx(a_reg[i], b_reg[i], result_reg[i])",
-            "        print(f'     CCX: {a_reg.name}[{i}] & {b_reg.name}[{i}] -> {result_reg.name}[{i}]')",
-            "",
-            "def apply_or_circuit(qc, a_reg, b_reg, result_reg):",
-            "    '''CORRECTED: Pure bitwise OR circuit'''",
-            "    print(f'   Applying pure OR circuit: {a_reg.name} | {b_reg.name} -> {result_reg.name}')",
-            "    # For bitwise OR: result[i] = a[i] | b[i]",
-            "    # Using: a | b = a ⊕ b ⊕ (a & b)",
-            "    for i in range(min(len(a_reg), len(b_reg), len(result_reg))):",
-            "        qc.cx(a_reg[i], result_reg[i])  # Copy a[i]",
-            "        qc.cx(b_reg[i], result_reg[i])  # XOR b[i]", 
-            "        qc.ccx(a_reg[i], b_reg[i], result_reg[i])  # XOR (a[i] & b[i])",
-            "        print(f'     OR bit {i}: {a_reg.name}[{i}] | {b_reg.name}[{i}] -> {result_reg.name}[{i}]')",
-            "",
-            "def apply_not_circuit(qc, input_reg, result_reg):",
-            "    '''CORRECTED: Pure bitwise NOT circuit'''",
-            "    print(f'   Applying pure NOT circuit: ~{input_reg.name} -> {result_reg.name}')",
-            "    # For bitwise NOT: result[i] = ~input[i]",
-            "    for i in range(min(len(input_reg), len(result_reg))):",
-            "        qc.cx(input_reg[i], result_reg[i])  # Copy input",
-            "        qc.x(result_reg[i])  # Flip bit",
-            "        print(f'     NOT bit {i}: ~{input_reg.name}[{i}] -> {result_reg.name}[{i}]')",
-            "",
-            "def apply_xor_circuit(qc, a_reg, b_reg, result_reg):",
-            "    '''XOR circuit'''",
-            "    for i in range(min(len(a_reg), len(b_reg), len(result_reg))):",
-            "        qc.cx(a_reg[i], result_reg[i])",
-            "        qc.cx(b_reg[i], result_reg[i])",
-            "",
-            "def apply_add_circuit(qc, a_reg, b_reg, result_reg):",
-            "    '''Addition circuit'''",
-            "    for i in range(min(len(a_reg), len(result_reg))):",
-            "        qc.cx(a_reg[i], result_reg[i])",
-            "    for i in range(min(len(b_reg), len(result_reg))):",
-            "        qc.cx(b_reg[i], result_reg[i])",
-            "    if len(result_reg) > 1 and len(a_reg) > 0 and len(b_reg) > 0:",
-            "        qc.ccx(a_reg[0], b_reg[0], result_reg[1])",
-            "",
-            "def apply_sub_circuit(qc, a_reg, b_reg, result_reg):",
-            "    '''Subtraction circuit'''",
-            "    for i in range(min(len(a_reg), len(result_reg))):",
-            "        qc.cx(a_reg[i], result_reg[i])",
-            "    for i in range(min(len(b_reg), len(result_reg))):",
-            "        qc.cx(b_reg[i], result_reg[i])",
-            "",
-            "def apply_mul_circuit(qc, a_reg, b_reg, result_reg):",
-            "    '''Multiplication circuit (simplified)'''",
-            "    for i in range(min(len(a_reg), len(result_reg))):",
-            "        if i < len(b_reg):",
-            "            qc.ccx(a_reg[i], b_reg[0], result_reg[i])",
-            "",
-            "def apply_div_circuit(qc, a_reg, b_reg, result_reg):",
-            "    '''Division circuit (simplified)'''",
-            "    if len(a_reg) > 1 and len(result_reg) > 0:",
-            "        qc.cx(a_reg[1], result_reg[0])",
-            "",
-            "def apply_mod_circuit(qc, a_reg, b_reg, result_reg):",
-            "    '''Modulo circuit'''",
-            "    for i in range(min(len(a_reg), len(result_reg))):",
-            "        qc.cx(a_reg[i], result_reg[i])",
-            "    if len(b_reg) > 0 and len(result_reg) > 0:",
-            "        qc.cx(b_reg[0], result_reg[0])",
-            "",
-            "def apply_gt_circuit(qc, a_reg, b_reg, result_reg):",
-            "    '''Greater than comparison circuit: result = (a > b)'''",
-            "    # Simplified greater than comparison",
-            "    # For 4-bit numbers, we implement a basic comparison",
-            "    # This is a simplified version - real quantum comparison is more complex",
-            "",
-            "    # Compare bit by bit from MSB to LSB",
-            "    # If a[i] = 1 and b[i] = 0 for any bit i, then a > b",
-            "    # We use ancilla qubits for intermediate results",
-            "",
-            "    # Simple implementation: check if a[1] > b[1] (MSB comparison)",
-            "    # This gives a reasonable approximation for the comparison",
-            "    if len(a_reg) > 1 and len(b_reg) > 1 and len(result_reg) > 0:",
-            "        # Create temporary ancilla for NOT b[1]",
-            "        qc.x(b_reg[1])  # Flip b[1] temporarily",
-            "        qc.ccx(a_reg[1], b_reg[1], result_reg[0])  # a[1] AND NOT b[1]",
-            "        qc.x(b_reg[1])  # Restore b[1]",
-            "    else:",
-            "        # Fallback: simple bit copy for basic comparison",
-            "        if len(a_reg) > 0 and len(result_reg) > 0:",
-            "            qc.cx(a_reg[0], result_reg[0])",
-            "",
-            "def apply_lt_circuit(qc, a_reg, b_reg, result_reg):",
-            "    '''Less than comparison circuit: result = (a < b)'''",
-            "    # a < b is equivalent to b > a",
-            "    apply_gt_circuit(qc, b_reg, a_reg, result_reg)",
-            "",
-            "def apply_eq_circuit(qc, a_reg, b_reg, result_reg):",
-            "    '''Equality comparison circuit: result = (a == b)'''",
-            "    # Simple equality check for LSB",
-            "    if len(a_reg) > 0 and len(b_reg) > 0 and len(result_reg) > 0:",
-            "        qc.cx(a_reg[0], result_reg[0])",
-            "        qc.cx(b_reg[0], result_reg[0])",
-            "        qc.x(result_reg[0])  # Flip to get equality (both same = 0 XOR = 1)",
-            "",
-            "def apply_ne_circuit(qc, a_reg, b_reg, result_reg):",
-            "    '''Not equal comparison circuit: result = (a != b)'''",
-            "    # XOR gives inequality",
-            "    if len(a_reg) > 0 and len(b_reg) > 0 and len(result_reg) > 0:",
-            "        qc.cx(a_reg[0], result_reg[0])",
-            "        qc.cx(b_reg[0], result_reg[0])",
-            "",
-            "def apply_neg_circuit(qc, input_reg, result_reg):",
-            "    '''Negation circuit'''",
-            "    for i in range(min(len(input_reg), len(result_reg))):",
-            "        qc.cx(input_reg[i], result_reg[i])",
-            "        qc.x(result_reg[i])",
-            "    qc.x(result_reg[0])",
-            "",
-            "def apply_post_inc_circuit(qc, input_reg, orig_reg, inc_reg):",
-            "    '''Post-increment circuit'''",
-            "    for i in range(min(len(input_reg), len(orig_reg))):",
-            "        qc.cx(input_reg[i], orig_reg[i])",
-            "    for i in range(min(len(input_reg), len(inc_reg))):",
-            "        qc.cx(input_reg[i], inc_reg[i])",
-            "    qc.x(inc_reg[0])",
-            "",
-            "def apply_post_dec_circuit(qc, input_reg, orig_reg, dec_reg):",
-            "    '''Post-decrement circuit'''",
-            "    for i in range(min(len(input_reg), len(orig_reg))):",
-            "        qc.cx(input_reg[i], orig_reg[i])",
-            "    for i in range(min(len(input_reg), len(dec_reg))):",
-            "        qc.cx(input_reg[i], dec_reg[i])",
-            "    qc.x(dec_reg[0])",
-            "",
-            "def create_quantum_circuit():",
-            "    '''Create the quantum circuit'''",
-            "    print('🔬 Creating quantum circuit...')",
-            "",
-        ]
-        
-        # Declare registers
-        declared_names = set()
-        for reg_name, reg_info in self.registers.items():
-            if reg_info.qiskit_name not in declared_names:
-                lines.append(f"    {reg_info.qiskit_name} = QuantumRegister({reg_info.size}, '{reg_info.qiskit_name}')")
-                declared_names.add(reg_info.qiskit_name)
-        
-        for cl_reg, qiskit_name in self.classical_registers.items():
-            if qiskit_name not in declared_names:
-                lines.append(f"    {qiskit_name} = ClassicalRegister(4, '{qiskit_name}')")
-                declared_names.add(qiskit_name)
-        
-        # Create circuit
-        unique_names = []
-        seen_names = set()
-        for reg_info in self.registers.values():
-            if reg_info.qiskit_name not in seen_names:
-                unique_names.append(reg_info.qiskit_name)
-                seen_names.add(reg_info.qiskit_name)
-        for cl_name in self.classical_registers.values():
-            if cl_name not in seen_names:
-                unique_names.append(cl_name)
-                seen_names.add(cl_name)
-
-        lines.append(f"    qc = QuantumCircuit({', '.join(unique_names)})")
-        lines.append("")
-        lines.append("    operations_log = []")
-        lines.append("")
-        
-        # Add operations
-        for i, op in enumerate(self.operations):
-            lines.append(f"    # Operation {i+1}: {op.description}")
-            
-            if op.op_type == "init":
-                reg_info = self._safe_get_register(op.operands[0])
-                if reg_info:
-                    value = op.attributes.get("value", 0)
-                    lines.append(f"    initialize_register(qc, {reg_info.qiskit_name}, {value}, {reg_info.size})")
-                    lines.append(f"    operations_log.append('Initialize {reg_info.qiskit_name} = {value}')")
-                    
-            elif op.op_type.endswith("_circuit"):
-                if len(op.operands) >= 3:
-                    reg_infos = [self._safe_get_register(operand) for operand in op.operands[:3]]
-                    if all(reg_infos):
-                        reg_names = [reg.qiskit_name for reg in reg_infos]
-                        lines.append(f"    apply_{op.op_type}(qc, {', '.join(reg_names)})")
-                        lines.append(f"    operations_log.append('{op.op_type}: {' + '.join(reg_names)}')")
-                elif len(op.operands) >= 2:
-                    reg_infos = [self._safe_get_register(operand) for operand in op.operands[:2]]
-                    if all(reg_infos):
-                        reg_names = [reg.qiskit_name for reg in reg_infos]
-                        lines.append(f"    apply_{op.op_type}(qc, {', '.join(reg_names)})")
-                        lines.append(f"    operations_log.append('{op.op_type}: {' -> '.join(reg_names)}')")
-                        
-            elif op.op_type in ["cx", "ccx", "x"]:
-                if op.op_type == "cx" and len(op.operands) >= 2:
-                    ctrl_reg, ctrl_idx = self._extract_qubit_index(op.operands[0])
-                    targ_reg, targ_idx = self._extract_qubit_index(op.operands[1])
-                    ctrl_info = self._safe_get_register(ctrl_reg)
-                    targ_info = self._safe_get_register(targ_reg)
-                    if ctrl_info and targ_info and ctrl_idx is not None and targ_idx is not None:
-                        lines.append(f"    qc.cx({ctrl_info.qiskit_name}[{ctrl_idx}], {targ_info.qiskit_name}[{targ_idx}])")
-                        lines.append(f"    operations_log.append('CNOT: {ctrl_info.qiskit_name}[{ctrl_idx}] -> {targ_info.qiskit_name}[{targ_idx}]')")
-                
-                elif op.op_type == "ccx" and len(op.operands) >= 3:
-                    ctrl1_reg, ctrl1_idx = self._extract_qubit_index(op.operands[0])
-                    ctrl2_reg, ctrl2_idx = self._extract_qubit_index(op.operands[1])
-                    targ_reg, targ_idx = self._extract_qubit_index(op.operands[2])
-                    ctrl1_info = self._safe_get_register(ctrl1_reg)
-                    ctrl2_info = self._safe_get_register(ctrl2_reg)
-                    targ_info = self._safe_get_register(targ_reg)
-                    if all([ctrl1_info, ctrl2_info, targ_info]) and all(idx is not None for idx in [ctrl1_idx, ctrl2_idx, targ_idx]):
-                        lines.append(f"    qc.ccx({ctrl1_info.qiskit_name}[{ctrl1_idx}], {ctrl2_info.qiskit_name}[{ctrl2_idx}], {targ_info.qiskit_name}[{targ_idx}])")
-                        lines.append(f"    operations_log.append('Toffoli: {ctrl1_info.qiskit_name}[{ctrl1_idx}] & {ctrl2_info.qiskit_name}[{ctrl2_idx}] -> {targ_info.qiskit_name}[{targ_idx}]')")
-                
-                elif op.op_type == "x" and len(op.operands) >= 1:
-                    targ_reg, targ_idx = self._extract_qubit_index(op.operands[0])
-                    targ_info = self._safe_get_register(targ_reg)
-                    if targ_info and targ_idx is not None:
-                        lines.append(f"    qc.x({targ_info.qiskit_name}[{targ_idx}])")
-                        lines.append(f"    operations_log.append('X gate: {targ_info.qiskit_name}[{targ_idx}]')")
-                        
-            elif op.op_type == "measure":
-                quantum_reg_info = self._safe_get_register(op.operands[0])
-                if quantum_reg_info and op.result in self.classical_registers:
-                    classical_reg = self.classical_registers[op.result]
-                    lines.append(f"    qc.measure({quantum_reg_info.qiskit_name}, {classical_reg})")
-                    lines.append(f"    operations_log.append('Measure: {quantum_reg_info.qiskit_name} -> {classical_reg}')")
-                    
-            lines.append("")
-        
-        # Complete the circuit function and add simulation
-        lines.extend([
-            "    return qc, operations_log",
-            "",
-            f"def run_simulation(qc, operations_log, expected_result={expected_result}):",
-            "    '''Run quantum simulation and analyze results'''",
-            "    print('🚀 Running quantum simulation...')",
-            "    print(f'Circuit: {qc.num_qubits} qubits, {qc.depth()} depth, {len(qc.data)} gates')",
-            "    print()",
-            "    ",
-            "    # Show operations",
-            "    print('📋 Quantum Operations:')",
-            "    for i, op in enumerate(operations_log, 1):",
-            "        print(f'  {i:2d}. {op}')",
-            "    print()",
-            "    ",
-            "    # Run simulation",
-            "    simulator = AerSimulator()",
-            "    job = simulator.run(qc, shots=1024)",
-            "    result = job.result()",
-            "    counts = result.get_counts()",
-            "    ",
-            "    # Analyze results",
-            "    print('📊 Measurement Results:')",
-            "    sorted_results = sorted(counts.items(), key=lambda x: x[1], reverse=True)",
-            "    ",
-            "    for bitstring, count in sorted_results[:10]:",
-            "        decimal = int(bitstring, 2)",
-            "        percentage = (count / 1024) * 100",
-            "        print(f'  {bitstring} (decimal: {decimal:2d}) -> {count:4d} shots ({percentage:5.1f}%)')",
-            "    ",
-            "    most_frequent_bits, most_frequent_count = sorted_results[0]",
-            "    quantum_result_unsigned = int(most_frequent_bits, 2)",
-            "    quantum_result_signed = convert_4bit_to_signed(quantum_result_unsigned)",
-            "    print(f'🎯 Quantum Result (unsigned): {quantum_result_unsigned}')",
-            "    print(f'🎯 Quantum Result (signed): {quantum_result_signed}')",
-            "    ",
-            "    print(f'🧮 Expected Result: {expected_result}')",
-            "    ",
-            "    if quantum_result_unsigned == expected_result:",
-            "        print('   ✅ PERFECT MATCH!')",
-            "        accuracy = 'PERFECT'",
-            "    else:",
-            "        difference = abs(quantum_result_unsigned - expected_result)",
-            "        print(f'   ⚠️  Difference: {difference}')",
-            "        ",
-            "        # Check if expected appears in results",
-            "        for bitstring, count in sorted_results:",
-            "            if int(bitstring, 2) == expected_result:",
-            "                percentage = (count / 1024) * 100",
-            "                print(f'   ✅ Expected result found with {percentage:.1f}% probability')",
-            "                break",
-            "        ",
-            "        accuracy = f'DIFF_{difference}'",
-            "    ",
-            "    return quantum_result_unsigned, accuracy, counts",
-            "",
-            "def visualize_circuit(qc):",
-            "    '''Show circuit diagram'''",
-            "    print('\\n📈 Circuit Visualization:')",
-            "    print(f'   Quantum qubits: {qc.num_qubits}')",
-            "    print(f'   Classical bits: {qc.num_clbits}')",
-            "    print(f'   Circuit depth: {qc.depth()}')",
-            "    print(f'   Total gates: {len(qc.data)}')",
-            "    ",
-            "    # Gate count",
-            "    gate_counts = {}",
-            "    for instruction in qc.data:",
-            "        gate_name = instruction.operation.name",
-            "        gate_counts[gate_name] = gate_counts.get(gate_name, 0) + 1",
-            "    ",
-            "    print('   Gate breakdown:')",
-            "    for gate, count in sorted(gate_counts.items()):",
-            "        print(f'     {gate}: {count}')",
-            "    ",
-            "    # Circuit diagram",
-            "    try:",
-            "        print('\\n   Circuit Diagram:')",
-            "        diagram = str(qc.draw(output='text', fold=-1))",
-            "        diagram_lines = diagram.split('\\n')",
-            "        for line in diagram_lines[:30]:  # Show first 30 lines",
-            "            print(f'   {line}')",
-            "        if len(diagram_lines) > 30:",
-            "            print(f'   ... ({len(diagram_lines) - 30} more lines)')",
-            "    except Exception as e:",
-            "        print(f'   Circuit diagram error: {e}')",
-            "",
-            "def main():",
-            "    '''Main execution function'''",
-            "    print('🚀 Quantum Circuit Execution')",
-            "    print('=' * 50)",
-            "    ",
-            "    try:",
-            "        # Create circuit",
-            "        print('📦 Creating quantum circuit...')",
-            "        qc, operations_log = create_quantum_circuit()",
-            "        ",
-            "        if not qc or not operations_log:",
-            "            print('❌ Failed to create circuit or operations log')",
-            "            return",
-            "        ",
-            "        # Visualize",
-            "        visualize_circuit(qc)",
-            "        ",
-            "        # Run simulation",
-            f"        quantum_result, accuracy, all_counts = run_simulation(qc, operations_log, {expected_result})",
-            "        ",
-            "        # Final summary",
-            "        print('\\n' + '=' * 50)",
-            "        print('🎊 Execution Complete!')",
-            "        print(f'   Quantum Result: {quantum_result}')",
-            f"        print(f'   Expected Result: {expected_result}')",
-            "        print(f'   Accuracy: {accuracy}')",
-            "        ",
-            "        return qc, operations_log, quantum_result",
-            "        ",
-            "    except Exception as e:",
-            "        print(f'❌ Error during execution: {e}')",
-            "        import traceback",
-            "        traceback.print_exc()",
-            "        return None, None, None",
-            "",
-            "# Execute immediately when script runs",
-            "print('🎬 STARTING QUANTUM CIRCUIT EXECUTION...')",
-            "print('='*60)",
-            "result = main()",
-            "if result and result[0]:",
-            "    print('\\n✅ QUANTUM CIRCUIT EXECUTION COMPLETED SUCCESSFULLY!')",
-            "else:",
-            "    print('\\n❌ QUANTUM CIRCUIT EXECUTION FAILED!')",
-            "",
-            "# Also ensure it runs with python -c or import",
-            "if __name__ == '__main__':",
-            "    if 'result' not in locals():",
-            "        print('\\n🔄 BACKUP EXECUTION...')",
-            "        backup_result = main()"
-        ])
-        
-        return lines
-    
-    def generate_complete_qiskit(self, mlir_content: str, expected_result: int) -> str:
-        """Generate complete Qiskit code from MLIR with external expected result"""
-        self.parse_optimized_mlir(mlir_content)
-        qiskit_lines = self.generate_qiskit_code(expected_result)
-        return '\n'.join(qiskit_lines)
-
-def main():
-    if len(sys.argv) < 3 or len(sys.argv) > 4:
-        print("Usage: python circuit_generator_fixed.py <optimized.mlir> <output.py> [expected_res.txt]")
-        print("\nFixed generator that:")
-        print("  ✅ Reads expected result from external file")
-        print("  ✅ Generates quantum circuits correctly")
-        print("  ✅ Shows accurate expected vs quantum results") 
-        print("  ✅ Circuit visualization and execution")
-        sys.exit(1)
-    
-    input_file = sys.argv[1]
-    output_file = sys.argv[2]
-    expected_file = sys.argv[3] if len(sys.argv) == 4 else "expected_res.txt"
-    
-    # Read MLIR
-    try:
-        with open(input_file, 'r') as f:
-            mlir_content = f.read()
-    except FileNotFoundError:
-        print(f"❌ Error: Input file '{input_file}' not found")
-        sys.exit(1)
-    
-    print(f"📄 Reading MLIR from: {input_file}")
-    
-    # Generate Qiskit code
-    generator = FixedQiskitGenerator()
-    
-    # Load expected result from external file
-    expected_result = generator.load_expected_result(expected_file)
-    
-    # Generate circuit with the loaded expected result
-    qiskit_code = generator.generate_complete_qiskit(mlir_content, expected_result)
-    
-    # Write output
-    try:
-        with open(output_file, 'w') as f:
-            f.write(qiskit_code)
-        print(f"✅ Generated circuit: {output_file}")
-    except Exception as e:
-        print(f"❌ Error writing output: {e}")
-        sys.exit(1)
-    
-    print(f"📊 {len(generator.registers)} registers, {len(generator.operations)} operations")
-    print(f"🧮 Expected result (from {expected_file}): {expected_result}")
-    print(f"🎯 Run with: python {output_file}")
-
-if __name__ == "__main__":
-    main()
